brandme-data/requirements.txt
# Copyright (c) Brand.Me, Inc. All rights reserved.

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Migrations
alembic==1.13.0
sqlalchemy==2.0.23

# CLI
click==8.1.7
python-dotenv==1.0.0
rich==13.7.0

# Utilities
pyyaml==6.0.1
